made a diff file for storage
